backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
asyncpg==0.29.0
alembic==1.13.3
python-multipart==0.0.12
pydantic==2.9.2
aiohttp==3.10.10
beautifulsoup4==4.12.3
lxml==5.3.0
openai==1.51.2
tavily-python==0.5.0
httpx==0.27.2
python-dotenv==1.0.1
